Add stage one install script
This script is what is returned by install.pi-hole.net, and will handle
the Pi-hole install. It will configure the package repository, install
the packages, make sure the services are configured correctly (disable
systemd-resolved), call the configuration script, and start the initial
gravity run.

The first two steps are implemented: configuring the package repository
and installing the packages. A testing repo is currently used, but will
be changed to the release repo when we are ready for release.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>